requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
litellm==1.30.0
python-dotenv==1.0.0
pandas==2.2.1 